Move debug settings to separate file
